add helper class and move DI and data seed there
